Rename v1/tv/ottv to v1/ao/opentracing
