+++ v1/ao/Makefile
@@ -1,8 +1,8 @@
-coverpkg="github.com/appoptics/appoptics-apm-go/v1/ao/internal/reporter,github.com/appoptics/appoptics-apm-go/v1/ao,github.com/appoptics/appoptics-apm-go/v1/ao/ottv"
+coverpkg="github.com/appoptics/appoptics-apm-go/v1/ao/internal/reporter,github.com/appoptics/appoptics-apm-go/v1/ao,github.com/appoptics/appoptics-apm-go/v1/ao/opentracing"
 cov_args=-covermode=atomic -coverpkg=$(coverpkg)
-cov_out=-coverprofile=cov.out 
-covtv_out=-coverprofile=covtv.out 
-cov_files=cov.out covtv.out internal/reporter/cov.out internal/reporter/covtv.out ottv/cov.out ottv/covtv.out
+cov_out=-coverprofile=cov.out
+covtv_out=-coverprofile=covao.out
+cov_files=cov.out covao.out internal/reporter/cov.out internal/reporter/covao.out opentracing/cov.out opentracing/covao.out
 cov_merge=covmerge.out
 test_args=-race
 
@@ -13,8 +13,8 @@ test:
 	go test $(test_args) $(cov_args) -v $(covtv_out) -tags disable_tracing
 	cd internal/reporter && go test $(test_args) $(cov_args) $(cov_out) -v
 	cd internal/reporter && go test $(test_args) $(cov_args) $(covtv_out) -v -tags disable_tracing
-	cd ottv && go test $(test_args) $(cov_args) $(cov_out) -v
-	cd ottv && go test $(test_args) $(cov_args) $(covtv_out) -v -tags disable_tracing
+	cd opentracing && go test $(test_args) $(cov_args) $(cov_out) -v
+	cd opentracing && go test $(test_args) $(cov_args) $(covtv_out) -v -tags disable_tracing
 
 $(cov_merge): test $(cov_files)
 	gocovmerge $(cov_files) > $(cov_merge)
