+++ v1/ao/http_client_instrumentation.go
@@ -11,7 +11,7 @@ import (
 
 // HTTPClientSpan is a Span that aids in reporting HTTP client requests.
 //   req, err := http.NewRequest("GET", "http://example.com", nil)
-//   l := tv.BeginHTTPClientSpan(ctx, httpReq)
+//   l := ao.BeginHTTPClientSpan(ctx, httpReq)
 //   defer l.End()
 //   // ...
 //   resp, err := client.Do(req)
