+++ v1/ao/opentracing/dapperish_test.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2017 Librato, Inc. All rights reserved.
 
-package ottv
+package opentracing
 
 import (
 	"bytes"
